feat: preload external images upfront
