+++ public/index.html
@@ -1,38 +1,43 @@
 <!DOCTYPE html>
 <html lang="en">
 
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#fafafa" />
-    <meta name="description" content="See what Genshin Matz are available today!" />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+<head>
+  <!-- Preload Images Upfront -->
+  <script src="%PUBLIC_URL%/preload.js"></script>
 
-    <!-- Google / Search Engine Tags -->
-    <meta itemprop="name" content="Genshin Matz Today">
-    <meta itemprop="description" content="See what Genshin Matz are available today!">
-    <meta itemprop="image" content="%PUBLIC_URL%/preview.png">
+  <title>Genshin Matz Today</title>
 
-    <!-- Facebook Meta Tags -->
-    <meta property="og:url" content="https://genshin-matz-today.netlify.app/">
-    <meta property="og:type" content="website">
-    <meta property="og:title" content="Genshin Matz Today">
-    <meta property="og:description" content="See what Genshin Matz are available today!">
-    <meta property="og:image" content="%PUBLIC_URL%/preview.png">
+  <meta charset="utf-8" />
+  <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <meta name="theme-color" content="#fafafa" />
+  <meta name="description" content="See what Genshin Matz are available today!" />
+  <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
+  <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
 
-    <!-- Twitter Meta Tags -->
-    <meta name="twitter:card" content="summary_large_image">
-    <meta name="twitter:title" content="Genshin Matz Today">
-    <meta name="twitter:description" content="See what Genshin Matz are available today!">
-    <meta name="twitter:image" content="%PUBLIC_URL%/preview.png">
-    <title>Genshin Matz Today</title>
-  </head>
+  <!-- Google / Search Engine Tags -->
+  <meta itemprop="name" content="Genshin Matz Today">
+  <meta itemprop="description" content="See what Genshin Matz are available today!">
+  <meta itemprop="image" content="%PUBLIC_URL%/preview.png">
 
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
-  </body>
+  <!-- Facebook Meta Tags -->
+  <meta property="og:url" content="https://genshin-matz-today.netlify.app/">
+  <meta property="og:type" content="website">
+  <meta property="og:title" content="Genshin Matz Today">
+  <meta property="og:description" content="See what Genshin Matz are available today!">
+  <meta property="og:image" content="%PUBLIC_URL%/preview.png">
+
+  <!-- Twitter Meta Tags -->
+  <meta name="twitter:card" content="summary_large_image">
+  <meta name="twitter:title" content="Genshin Matz Today">
+  <meta name="twitter:description" content="See what Genshin Matz are available today!">
+  <meta name="twitter:image" content="%PUBLIC_URL%/preview.png">
+
+</head>
+
+<body>
+  <noscript>You need to enable JavaScript to run this app.</noscript>
+  <div id="root"></div>
+</body>
 
 </html>
