+++ src/data/items.js
@@ -1,0 +1,80 @@
+const items = {
+  "Talent": [
+    {
+      name: "Ballad Series",
+      schedule: ["Wednesday", "Saturday", "Sunday"],
+      location: "FORSAKEN RIFT",
+      image: "https://vignette.wikia.nocookie.net/gensin-impact/images/4/46/Item_Philosophies_of_%22Ballad%22.png/revision/latest?cb=20200715011445"
+    },
+    {
+      name: "Diligence Series",
+      schedule: ["Tuesday", "Friday", "Sunday"],
+      location: "TAISHAN MANSION",
+      image: "https://vignette.wikia.nocookie.net/gensin-impact/images/d/d5/Item_Philosophies_of_%22Diligence%22.png/revision/latest?cb=20200715011548"
+    },
+    {
+      name: "Freedom Series",
+      schedule: ["Monday", "Thursday", "Sunday"],
+      location: "FORSAKEN RIFT",
+      image: "https://vignette.wikia.nocookie.net/gensin-impact/images/4/47/Item_Philosophies_of_%22Freedom%22.png/revision/latest?cb=20200715011355"
+    },
+    {
+      name: "Gold Series",
+      schedule: ["Wednesday", "Saturday", "Sunday"],
+      location: "TAISHAN MANSION",
+      image: "https://vignette.wikia.nocookie.net/gensin-impact/images/b/b4/Item_Philosophies_of_%22Gold%22.png/revision/latest?cb=20200715011655"
+    },
+    {
+      name: "Prosperity Series",
+      schedule: ["Monday", "Thursday", "Sunday"],
+      location: "TAISHAN MANSION",
+      image: "https://vignette.wikia.nocookie.net/gensin-impact/images/9/98/Item_Philosophies_of_%22Prosperity%22.png/revision/latest?cb=20200715011515"
+    },
+    {
+      name: "Resistance Series",
+      schedule: ["Tuesday", "Friday", "Sunday"],
+      location: "FORSAKEN RIFT",
+      image: "https://vignette.wikia.nocookie.net/gensin-impact/images/1/14/Item_Philosophies_of_%22Resistance%22.png/revision/latest?cb=20200715011426"
+    },
+  ],
+  "Weapon Ascension": [
+    {
+      name: "Boreal Wolf Series",
+      schedule: ["Tuesday", "Friday", "Sunday"],
+      location: "CECILIA GARDEN",
+      image: "https://vignette.wikia.nocookie.net/gensin-impact/images/e/e0/Item_Boreal_Wolf%27s_Nostalgia.png/revision/latest?cb=20200711153726"
+    },
+    {
+      name: "Decarabian Series",
+      schedule: ["Monday", "Thursday", "Sunday"],
+      location: "CECILIA GARDEN",
+      image: "https://vignette.wikia.nocookie.net/gensin-impact/images/4/4c/Item_Scattered_Piece_of_Decarabian%27s_Dream.png/revision/latest?cb=20200711151726"
+    },
+    {
+      name: "Dandelion Gladiator Series",
+      schedule: ["Wednesday", "Saturday", "Sunday"],
+      location: "CECILIA GARDEN",
+      image: "https://vignette.wikia.nocookie.net/gensin-impact/images/b/b4/Item_Dream_of_the_Dandelion_Gladiator.png/revision/latest?cb=20200711160234"
+    },
+    {
+      name: "Guyun Series",
+      schedule: ["Monday", "Thursday", "Sunday"],
+      location: "HIDDEN PALACE OF LIANSHAN FORMULA",
+      image: "https://vignette.wikia.nocookie.net/gensin-impact/images/e/e8/Item_Divine_Body_from_Guyun.png/revision/latest?cb=20200711161335"
+    },
+    {
+      name: "Mist Veiled Elixir Series",
+      schedule: ["Tuesday", "Friday", "Sunday"],
+      location: "HIDDEN PALACE OF LIANSHAN FORMULA",
+      image: "https://vignette.wikia.nocookie.net/gensin-impact/images/a/a3/Item_Mist_Veiled_Primo_Elixir.png/revision/latest?cb=20200711163944"
+    },
+    {
+      name: "Aerosiderite Series",
+      schedule: ["Wednesday", "Saturday", "Sunday"],
+      location: "HIDDEN PALACE OF LIANSHAN FORMULA",
+      image: "https://vignette.wikia.nocookie.net/gensin-impact/images/4/48/Item_Chunk_of_Aerosiderite.png/revision/latest?cb=20200711164843"
+    },
+  ]
+}
+
+export default items
